requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
starlette==0.27.0
fastapi[all]
python-dotenv==1.0.0